Merge

+++ airflow-core/src/airflow/utils/db_manager.py
@@ -23,6 +23,7 @@ from alembic import command
 from sqlalchemy import inspect
 
 from airflow import settings
+from airflow.api_fastapi.app import create_auth_manager
 from airflow.configuration import conf
 from airflow.exceptions import AirflowException
 from airflow.utils.log.logging_mixin import LoggingMixin
@@ -42,6 +43,7 @@ class BaseDBManager(LoggingMixin):
     version_table_name: str
     # Whether the database supports dropping tables when airflow tables are dropped
     supports_table_dropping: bool = False
+    revision_heads_map: dict[str, str] = {}
 
     def __init__(self, session):
         super().__init__()
@@ -96,6 +98,8 @@ class BaseDBManager(LoggingMixin):
         self.log.info("%s tables have been created from the ORM", self.__class__.__name__)
 
     def drop_tables(self, connection):
+        if not self.supports_table_dropping:
+            return
         self.metadata.drop_all(connection)
         version = self._get_migration_ctx()._version
         if inspect(connection).has_table(version.name):
@@ -143,7 +147,15 @@ class RunDBManager(LoggingMixin):
     def __init__(self):
         super().__init__()
         self._managers: list[BaseDBManager] = []
-        managers = conf.get("database", "external_db_managers").split(",")
+        managers_config = conf.get("database", "external_db_managers", fallback=None)
+        if not managers_config:
+            managers = []
+        else:
+            managers = managers_config.split(",")
+        # Add DB manager specified by auth manager (if any)
+        auth_manager_db_manager = create_auth_manager().get_db_manager()
+        if auth_manager_db_manager and auth_manager_db_manager not in managers:
+            managers.append(auth_manager_db_manager)
         for module in managers:
             manager = import_string(module)
             self._managers.append(manager)
@@ -213,6 +225,5 @@ class RunDBManager(LoggingMixin):
     def drop_tables(self, session, connection):
         """Drop the external database managers."""
         for manager in self._managers:
-            if manager.supports_table_dropping:
-                m = manager(session)
-                m.drop_tables(connection)
+            m = manager(session)
+            m.drop_tables(connection)
